pyproject.toml
[tool.poetry]
name = "ml-service"
version = "0.1.0"
description = ""
authors = ["Egor Bodrov <bodroff.e.a@gmail.com>"]
packages = [{include = "src"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
sqlalchemy = "^2.0.40"
asyncpg = "^0.30.0"
pydantic-settings = "^2.9.1"
python-dotenv = "^1.1.0"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
rq = "^2.3.3"


[tool.poetry.group.ml.dependencies]
scikit-learn = "^1.6.1"
pandas = "^2.2.3"
numpy = "^2.2.3"
matplotlib = "^3.10.1"
seaborn = "^0.13.2"


[tool.poetry.group.web.dependencies]
streamlit = "^1.45.0"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
